backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
openai==1.99.9
python-dotenv==1.1.1
